+++ shell/ls.go
@@ -1,13 +1,32 @@
 package shell
 
-import 	"github.com/abiosoft/ishell"
+import (
+	"errors"
+
+	"github.com/abiosoft/ishell"
+)
 
 func lsCmd(ctx *ShellCtxt) *ishell.Cmd {
 	return &ishell.Cmd{
 		Name: "ls",
-		Help: "list current directory",
+		Help: "list directory",
 		Func: func(c *ishell.Context) {
-			for _, e := range ctx.node.Children {
+			node := ctx.node
+
+			if len(c.Args) == 1 {
+				target := c.Args[0]
+
+				argNode, err := ctx.api.Filetree.NodeByPath(target, ctx.node)
+
+				if err != nil || node.IsFile() {
+					c.Err(errors.New("directory doesn't exist"))
+					return
+				}
+
+				node = argNode
+			}
+
+			for _, e := range node.Children {
 				eType := "d"
 				if e.IsFile() {
 					eType = "f"